git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1081 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/G3D.lib/include/G3D/PointHashGrid.h
@@ -1023,7 +1023,7 @@ public:
     }
 
     int debugGetDeepestBucketSize() const {
-        return m_data.debugGetDeepestBucketSize();
+        return (int)m_data.debugGetDeepestBucketSize();
     }
 
     float debugGetAverageBucketSize() const {
